Refactor non-distributed automl utils into a separate directory.
