+++ ludwig/automl/__init__.py
@@ -1,0 +1,1 @@
+from ludwig.automl.automl import auto_train, cli_init_config, create_auto_config, train_with_config  # noqa
