+++ ludwig/automl/base_config.py
@@ -20,11 +20,11 @@ import numpy as np
 import pandas as pd
 from dataclasses_json import dataclass_json, LetterCase
 
-from ludwig.automl.data_source import DataframeSource, DataSource
-from ludwig.automl.field_info import FieldConfig, FieldInfo, FieldMetadata
-from ludwig.automl.ray_utils import _ray_init, get_available_resources
-from ludwig.automl.type_inference import infer_type, should_exclude
 from ludwig.constants import COMBINER, EXECUTOR, HYPEROPT, SCHEDULER, SEARCH_ALG, TEXT, TYPE
+from ludwig.utils.automl.data_source import DataframeSource, DataSource
+from ludwig.utils.automl.field_info import FieldConfig, FieldInfo, FieldMetadata
+from ludwig.utils.automl.ray_utils import _ray_init, get_available_resources
+from ludwig.utils.automl.type_inference import infer_type, should_exclude
 from ludwig.utils.data_utils import load_dataset, load_yaml
 from ludwig.utils.defaults import default_random_seed
