+++ tests/ludwig/utils/automl/test_type_inference.py
@@ -2,10 +2,10 @@ import random
 
 import pytest
 
-from ludwig.automl.field_info import FieldInfo
-from ludwig.automl.type_inference import infer_type, should_exclude
 from ludwig.constants import AUDIO, BINARY, CATEGORY, DATE, IMAGE, NUMBER, TEXT
 from ludwig.data.dataset_synthesizer import generate_string
+from ludwig.utils.automl.field_info import FieldInfo
+from ludwig.utils.automl.type_inference import infer_type, should_exclude
 
 ROW_COUNT = 100
 TARGET_NAME = "target"
